backend/requirements.txt
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
pip==24.3.1
psycopg2==2.9.10
pillow==11.0.0
drf-spectacular==0.28.0